move functions in const to functions.js, persist chat state in theater mode
